+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `get-contexts-by-selectors` to get a subset of the existing contexts.
+
 ## [0.1.0-alpha-6] - 2024.10.01
 
 ### Fixed